🧬 Migrate the data model *BaseConfig* to module *api_config.base*.
