+++ test/unit_test/model/api_config/base.py
@@ -1,0 +1,26 @@
+import pytest
+
+from pymock_api.model.api_config import BaseConfig
+
+from ...._values import _Base_URL, _TestConfig
+from ._base import ConfigTestSpec, _assertion_msg
+
+
+class TestBaseConfig(ConfigTestSpec):
+    @pytest.fixture(scope="function")
+    def sut(self) -> BaseConfig:
+        return BaseConfig(url=_Base_URL)
+
+    @pytest.fixture(scope="function")
+    def sut_with_nothing(self) -> BaseConfig:
+        return BaseConfig()
+
+    def test_value_attributes(self, sut: BaseConfig):
+        assert sut.url == _Base_URL, _assertion_msg
+
+    def _expected_serialize_value(self) -> dict:
+        return _TestConfig.Base
+
+    def _expected_deserialize_value(self, obj: BaseConfig) -> None:
+        assert isinstance(obj, BaseConfig)
+        assert obj.url == _Base_URL
